refactor: 주문하기 기능을 커스텀 훅으로 분리
