+++ src/pages/OrderPage/OrderPage.tsx
@@ -1,66 +1,48 @@
 import styled from '@emotion/styled'
-import { useState, useMemo } from 'react'
 import { Navbar } from '@/components/Navbar/Navbar'
 import { cardMock } from '@/pages/OrderPage/cardMock'
 import { Layout } from '@/components/Layout/Layout'
 import { useParams } from 'react-router-dom'
 import { productMock } from '@/components/Product/productMock'
+import { useOrderForm } from '@/hooks/useOrderForm'
+import { useMemo } from 'react'
 
 export function OrderPage() {
-  const [selectedCard, setSelectedCard] = useState(cardMock[0])
-  const [message, setMessage] = useState(selectedCard.defaultTextMessage)
-  const [sender, setSender] = useState('')
-  const [receiver, setReceiver] = useState('')
-  const [receiverPhone, setReceiverPhone] = useState('')
-  const [quantity, setQuantity] = useState(1)
-  const [isSubmitted, setIsSubmitted] = useState(false)
-
   const { id } = useParams<{ id: string }>()
   const product = useMemo(
     () => productMock.find((p) => p.id === Number(id)),
     [id]
   )
 
+  const {
+    selectedCard,
+    setSelectedCard,
+    message,
+    setMessage,
+    sender,
+    setSender,
+    receiver,
+    setReceiver,
+    receiverPhone,
+    setReceiverPhone,
+    quantity,
+    setQuantity,
+    isSubmitted,
+    setIsSubmitted,
+    senderError,
+    receiverError,
+    receiverPhoneError,
+    messageError,
+    quantityError,
+    isFormValid,
+    resetForm,
+  } = useOrderForm()
+
   if (!product) return <div>상품을 찾을 수 없습니다.</div>
 
   const price = product.price.sellingPrice
   const productName = product.name
   const brandName = product.brandInfo.name
-
-  const senderError = useMemo(() => {
-    if (!sender) return '보내는 사람 이름을 입력해주세요.'
-    return ''
-  }, [sender])
-
-  const receiverError = useMemo(() => {
-    if (!receiver) return '받는 사람 이름을 입력해주세요.'
-    return ''
-  }, [receiver])
-
-  const phoneRegex = /^010\d{8}$/
-  const receiverPhoneError = useMemo(() => {
-    if (!receiverPhone) return '받는 사람 전화번호를 입력해주세요.'
-    if (!phoneRegex.test(receiverPhone))
-      return '올바른 전화번호 형식이 아닙니다.'
-    return ''
-  }, [receiverPhone])
-
-  const messageError = useMemo(() => {
-    if (!message) return '메시지를 입력해주세요.'
-    return ''
-  }, [message])
-
-  const quantityError = useMemo(() => {
-    if (quantity < 1) return '수량은 1개 이상이어야 합니다.'
-    return ''
-  }, [quantity])
-
-  const isFormValid =
-    !senderError &&
-    !receiverError &&
-    !receiverPhoneError &&
-    !messageError &&
-    !quantityError
 
   const submitOrderForm = (e: React.FormEvent) => {
     e.preventDefault()
@@ -72,13 +54,7 @@ export function OrderPage() {
         구매 수량: ${quantity}
         발신자 이름: ${sender}
         메시지: ${message}`)
-
-      setSender('')
-      setReceiver('')
-      setReceiverPhone('')
-      setMessage('')
-      setQuantity(1)
-      setIsSubmitted(false)
+      resetForm()
     }
   }
